requirements.txt
aiohttp==3.7.1
async-timeout==3.0.1
attrs==20.2.0
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
decorator==4.4.2
fsspec==0.8.4
gcsfs==0.7.1
google-auth==1.22.1
google-auth-oauthlib==0.4.1
idna==2.10
multidict==5.0.0
numpy==1.19.2
oauthlib==3.1.0
pandas==1.1.3
providah==0.1.15.0
pyarrow==2.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
six==1.15.0
typing-extensions==3.7.4.3
urllib3==1.25.11
yarl==1.6.2
